Simplify and reduce the number of test files no code changes
